requirements.txt
gitpython
click
pyyaml
pathlib